# mime: reduce allocs incurred by ParseMediatype